+++ src/platform/linux-mips/mod.rs
@@ -96,4 +96,42 @@ pub unsafe fn syscall4(mut nr: usize,
     if a4 == 0 { nr } else { -(nr as isize) as usize }
 }
 
-// TODO syscall{5,6} arguments are passed through the stack
+#[inline(always)]
+pub unsafe fn syscall5(mut nr: usize,
+                       a1: usize,
+                       a2: usize,
+                       a3: usize,
+                       mut a4: usize,
+                       a5: usize)
+                       -> usize {
+    asm!("subu $$29,20
+          sw $5, 16($$29)
+          syscall
+          addiu $$29,20"
+         : "+{$2}"(nr) "+{$7}"(a4)
+         : "{$4}"(a1) "{$5}"(a2) "{$6}"(a3) "r"(a5)
+         : "$8" "$9" "$10" "$11" "$12" "$13" "$14" "$15" "$24" "$25" "memory"
+         : "volatile");
+    if a4 == 0 { nr } else { -(nr as isize) as usize }
+}
+
+#[inline(always)]
+pub unsafe fn syscall6(mut nr: usize,
+                       a1: usize,
+                       a2: usize,
+                       a3: usize,
+                       mut a4: usize,
+                       a5: usize,
+                       a6: usize)
+                       -> usize {
+    asm!("subu $$29,24
+          sw $5, 16($$29)
+          sw $6, 20($$29)
+          syscall
+          addiu $$29,24"
+         : "+{$2}"(nr) "+{$7}"(a4)
+         : "{$4}"(a1) "{$5}"(a2) "{$6}"(a3) "r"(a5) "r"(a6)
+         : "$8" "$9" "$10" "$11" "$12" "$13" "$14" "$15" "$24" "$25" "memory"
+         : "volatile");
+    if a4 == 0 { nr } else { -(nr as isize) as usize }
+}